# Docs/OpenReferralApiOpenApiSpec.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Open Referrals API(Cast Project)",
    "description": "Allows users to query Orgs, Services and Locations. Data can be added easily using the related OpenReferralUI project to sign up and add data.",
    "contact": {
      "name": "Catalyst",
      "url": "https://www.thecatalyst.org.uk/",
      "email": "support@wearecast.org.uk"
    },
    "version": "v1"
  },
  "paths": {
...
    "/Services": {
      "get": {
        "tags": [
          "Services"
        ],
        "summary": "Get All Services",
        "parameters": [
          {
            "name": "postcode",
            "in": "query",
            "description": "Use text to perform a keyword search on services. This performs a full text search on the service title.",
            "schema": {
              "type": "string",
              "description": "Use text to perform a keyword search on services. This performs a full text search on the service title.",
              "nullable": true
            }
          },
          {
            "name": "proximity",
            "in": "query",
            "description": "The distance in metres that the person is willing to travel from the target postcode.",
            "schema": {
              "type": "number",
              "description": "The distance in metres that the person is willing to travel from the target postcode.",
              "format": "double",
              "nullable": true
            }
          },
          {
            "name": "text",
            "in": "query",
            "description": "The postcode of the person who wishes to use the service. In order to find services that are within a reasonable distance.",
            "schema": {
              "type": "string",
              "description": "The postcode of the person who wishes to use the service. In order to find services that are within a reasonable distance.",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "Services"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Service"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Service"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Service"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Services/{id}": {
      "get": {
        "tags": [
          "Services"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Services"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Service"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Service"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Service"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ServiceAtLocation": {
        "required": [
          "id",
          "location_Id",
          "service_Id"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "location_Id": {
            "type": "string"
          },
          "service_Id": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "RegularSchedule": {
        "type": "object",
        "properties": {
          "byDay": {
            "type": "string",
            "nullable": true
          },
          "byMonthDay": {
            "type": "string",
            "nullable": true
          },
          "closes_At": {
            "type": "string",
            "format": "date-time"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "dtStart": {
            "type": "string",
            "nullable": true
          },
          "freq": {
            "type": "string",
            "nullable": true
          },
          "id": {
            "type": "string",
            "nullable": true
          },
          "interval": {
            "type": "string",
            "nullable": true
          },
          "opens_At": {
            "type": "string",
            "format": "date-time"
          },
          "service": {
            "$ref": "#/components/schemas/Service"
          },
          "service_At_Location": {
            "$ref": "#/components/schemas/ServiceAtLocation"
          },
          "valid_From": {
            "type": "string",
            "format": "date-time"
          },
          "valid_To": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "Service": {
        "required": [
          "description",
          "email",
          "id",
          "name",
          "url"
        ],
        "type": "object",
        "properties": {
          "description": {
            "maxLength": 200,
            "minLength": 0,
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "id": {
            "type": "string"
          },
          "name": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string"
          },
          "organizationId": {
            "type": "string",
            "nullable": true
          },
          "regular_Schedules": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RegularSchedule"
            },
            "nullable": true
          },
          "service_At_Locations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ServiceAtLocation"
            },
            "nullable": true
          },
          "contacts": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "status": {
            "type": "string",
            "nullable": true
          },
          "url": {
            "type": "string",
            "format": "uri"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          },
          "category": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "Note that a token can only be created via the correct AzureB2C tenant. JWT Authorization header using the Bearer scheme. Enter 'Bearer' [space] and then your token in the text input below. Example: 'Bearer 12345abcdef'",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": []
    }
  ]
}